requirements.txt
asgiref==3.5.2
distlib==0.3.6
Django==4.1.1
django-filter==22.1
djangorestframework==3.13.1
filelock==3.8.0
platformdirs==2.5.2
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
virtualenv==20.16.5
